+++ src/js/core.js
@@ -43,8 +43,12 @@ $(document).ready(function ($) {
                 // Add handling for different kinds of events. For ex: {"event": "notification", "data": {"title": ..}}
                 if (msg.event === 'notification') {
                     SICKRAGE.notify(msg.data.type, msg.data.title, msg.data.body);
-                } else if (msg.event === 'redirect') {
-                    window.location.href = msg.data.url;
+                } else if (msg.event === 'task') {
+                    switch (msg.data.cmd) {
+                        case 'restart':
+                            window.location.href = SICKRAGE.srWebRoot + '/home/restart/?pid=' + SICKRAGE.srPID;
+                            break;
+                    }
                 }
             };
         },